+++ frontend/src/infrastructure.ts
@@ -13,6 +13,7 @@ export default class Infrastructure {
     _data: InfrastructureData;  // Unlike other data classes, Infra data will not hold the modules/connectors lists themselves, but will be passed data about their coordinates, etc so that it can perform checks on potential locations' validity
     _modules: Module[]; 
     _connectors: Connector[];
+    _currentSerial: number;     // A crude but hopefully effective method of ID-ing structures as they're created
     _horizontalOffset: number;  // Value is in pixels
 
     // Map width is passed to the data class at construction to help with base volume calculations
@@ -21,11 +22,38 @@ export default class Infrastructure {
         this._data = new InfrastructureData();
         this._modules = [];
         this._connectors = [];
+        this._currentSerial = 1000;     // If changing, change the value in the reset method too
         this._horizontalOffset = 0;
     }
 
     setup(mapWidth: number) {
         this._data.setup(mapWidth)
+    }
+
+    addModule (x: number, y: number, moduleInfo: ModuleInfo) {
+        const m = new Module(this._p5, this._currentSerial, x, y, moduleInfo);
+        this._modules.push(m);
+        this.increaseSerialNumber();
+        // Update base volume data
+        const area = this._data.calculateModuleArea(moduleInfo, x, y);
+        this._data.updateBaseVolume(area);
+        // Update base floor data
+        const footprint = this._data.calculateModuleFootprint(area);
+        // this._data.addModuleToFloors(moduleInfo, footprint);
+    }
+
+    addConnector (start: Coords, stop: Coords, connectorInfo: ConnectorInfo) {
+        const c = new Connector(this._p5, this._currentSerial, start, stop, connectorInfo)
+        this._connectors.push(c);
+        this.increaseSerialNumber();
+    }
+
+    increaseSerialNumber () {
+        this._currentSerial++;
+    }
+
+    resetSerialNumber () {
+        this._currentSerial = 1000;
     }
 
     // Top level module placement checker: Calls sub-routines from the data class
@@ -52,18 +80,6 @@ export default class Infrastructure {
             // console.log(`Module gaps underneath module: ${modFloor}`);
             return false;
         }
-    }
-
-    addModule (x: number, y: number, moduleInfo: ModuleInfo) {
-
-        this._modules.push(new Module(this._p5, x, y, moduleInfo));
-        // Update base volume data
-        const area = this._data.calculateModuleArea(moduleInfo, x, y);
-        this._data.updateBaseVolume(area);
-    }
-
-    addConnector (start: Coords, stop: Coords, connectorInfo: ConnectorInfo) {
-        this._connectors.push(new Connector(this._p5, start, stop, connectorInfo));
     }
 
     //  Takes in data for a new module's location and compares it to all of the other existing modules to look for overlaps